backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1
pymongo==4.6.0
